Setting up infrastructure for kinetic solvers. Ported over most of the files from the earlier MOOSE, still need to complete the implementation of KineticHub. It all compiles but no tests yet.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@96 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 7faa18007b5a8b89742a505811799656d232e698

+++ kinetics/Molecule.cpp
@@ -175,7 +175,7 @@ void Molecule::setMode( const Conn& c, int value )
 	static_cast< Molecule* >( c.data() )->mode_ = value;
 }
 
-double Molecule::getMode( const Element* e )
+int Molecule::getMode( const Element* e )
 {
 	return static_cast< Molecule* >( e->data() )->mode_;
 }
